requirements.txt
streamlit
supabase
easyocr
pymupdf
pandas
pillow
numpy
opencv-python-headless